Deduplicate IDE modules based on available projects
Instead of de-duplication based on post-processing the configured
IDE module names, de-duplication is now done for the default IDE
module names, based on the set of available projects. This results
in slightly more predictable behaviour, since the default name of
a module is not dependent on which projects have the IDE plugin
applied.

In addition, we now _always_ honour the user-configured IDE module
name, even if this will result in a duplicate name.